endpoints para proveedores y cambios en users

+++ backend/model/userModel.js
@@ -11,11 +11,11 @@ async function getAllUsers() {
   }
 }
 
-async function createUser(email, contrasena, rol) {
+async function createUser(email, password, rol) {
   try {
     const res = await pool.query(
       "INSERT INTO Usuarios (email, contrasena, rol) VALUES ($1, $2, $3) RETURNING usuario_id",
-      [email, contrasena, rol]
+      [email, password, rol]
     );
     return res.rows[0];
   } catch (error) {
@@ -34,11 +34,11 @@ async function getUserByEmail(email) {
   }
 }
 
-async function getUserById(usuario_id) {
+async function getUserById(user_id) {
   try {
     const res = await pool.query(
       "SELECT * FROM Usuarios WHERE usuario_id = $1",
-      [usuario_id]
+      [user_id]
     );
 
     if (res.rows.length === 0) {
@@ -51,26 +51,24 @@ async function getUserById(usuario_id) {
   }
 }
 
-async function updateUser(usuario_id, email, contrasena, rol) {
+async function updateUser(user_id, email, password, rol) {
   try {
     const res = await pool.query(
       `UPDATE Usuarios 
        SET email = $1, contrasena = $2, rol = $3 
        WHERE usuario_id = $4 
        RETURNING *`,
-      [email, contrasena, rol, usuario_id]
+      [email, password, rol, user_id]
     );
-    return res.rows[0]; 
+    return res.rows[0];
   } catch (error) {
     throw error;
   }
 }
 
-async function deleteUser(usuario_id) {
+async function deleteUser(user_id) {
   try {
-    await pool.query("DELETE FROM Usuarios WHERE usuario_id = $1", [
-      usuario_id,
-    ]);
+    await pool.query("DELETE FROM Usuarios WHERE usuario_id = $1", [user_id]);
   } catch (error) {
     throw error;
   }
